License under GPLv3 instead of BSD-3 (ugh)
Steinberg only lets VST3 plugins be licensed under their proprietary license (ew) or the GPLv3 (ew but less ew). Because I want to keep this code free, it's licensed under the GPLv3. It would be cool if the BSD-3 were a valid license for VST3 plugins though. Steinberg, take notes!